pyproject.toml
[tool.poetry]
name = "eventmap"
version = "0.1.0"
description = ""
authors = ["Vasilii Safronov <vasi1988@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-login = "^0.6.2"
flask-wtf = "^1.1.1"
psycopg2-binary = "^2.9.6"
pymongo = "^4.4.0"
requests = "^2.31.0"
email-validator = "^2.0.0.post2"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.5"
geoalchemy2 = "^0.13.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
